+++ packages/ice-urql/src/index.ts
@@ -1,8 +1,9 @@
 import type { Plugin } from '@ice/app/types';
 
 const PLUGIN_NAME = '@knockout-js/ice-urql';
-
-const plugin: Plugin = () => ({
+interface PluginIceUrqlOptions {
+}
+const plugin: Plugin<PluginIceUrqlOptions | void> = () => ({
   name: PLUGIN_NAME,
   setup: () => {
   },