+++ packages/layout/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @knockout-js/layout
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @knockout-js/ice-urql@0.1.6
+
 ## 0.1.7
 
 ### Patch Changes